# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TsoaResponse_string_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"type": "string"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UUID": {
				"type": "string",
				"pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
			},
			"SimplePostResponseDTO": {
				"description": "리소스 생성 성공 시 반환되는 기본 응답 객체",
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID",
						"format": "uuid"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TsoaResponse_SimplePostResponseDTO_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/SimplePostResponseDTO"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateChatRoomDTO": {
				"description": "채팅방 생성 요청 DTO",
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID",
						"description": "요청글, 제안서, 프로필등 채팅방을 생성하는 주체의 고유 아이디"
					},
					"type": {
						"type": "string",
						"enum": [
							"REQUEST",
							"PROPOSAL",
							"FEED"
						]
					},
					"myId": {
						"$ref": "#/components/schemas/UUID",
						"description": "채팅방을 생성하는 사용자 ID (테스트용 - 로그인 구현 전까지 사용)"
					}
				},
				"required": [
					"id",
					"type",
					"myId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChatRoomPreviewDTO": {
				"properties": {
					"chatRoomId": {
						"$ref": "#/components/schemas/UUID",
						"example": "123e4567-e89b-12d3-a456-426614174000",
						"format": "uuid"
					},
					"image": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"roomType": {
						"type": "string"
					},
					"messageType": {
						"type": "string",
						"enum": [
							"TEXT",
							"IMAGE",
							"OTHER"
						]
					},
					"type": {
						"type": "string",
						"enum": [
							"INQUIRY",
							"ORDER"
						]
					},
					"lastMessage": {
						"$ref": "#/components/schemas/UUID"
					},
					"lastMessageAt": {
						"type": "string",
						"format": "date-time"
					},
					"unreadCount": {
						"type": "number",
						"format": "double",
						"description": "안 읽은 메시지 수",
						"example": 3,
						"minimum": 0
					}
				},
				"required": [
					"chatRoomId",
					"image",
					"title",
					"roomType",
					"messageType",
					"type",
					"lastMessage",
					"lastMessageAt",
					"unreadCount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pagenation": {
				"properties": {
					"hasMore": {
						"type": "boolean"
					},
					"nextCursor": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UUID"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"hasMore",
					"nextCursor"
				],
				"type": "object",
				"description": "페이지네이션 메타데이터"
			},
			"ChatRoomListDTO": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ChatRoomPreviewDTO"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Pagenation"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TsoaResponse_ChatRoomListDTO_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/ChatRoomListDTO"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChatRoomFilter": {
				"type": "string",
				"enum": [
					"INQUIRY",
					"ORDER",
					"UNREAD"
				]
			},
			"SendSmsResponse": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"statusCode",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TsoaResponse_SendSmsResponse_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/SendSmsResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"properties": {
							"data": {
								"nullable": true
							},
							"reason": {
								"type": "string",
								"nullable": true
							},
							"errorCode": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"success": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendSmsRequest": {
				"properties": {
					"phoneNumber": {
						"type": "string",
						"description": "인증번호를 받을 휴대폰 번호",
						"example": "01012345678",
						"pattern": "^01[016789]-?\\d{3,4}-?\\d{4}$",
						"minLength": 10,
						"maxLength": 13
					}
				},
				"required": [
					"phoneNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifySmsResponse": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"statusCode",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TsoaResponse_VerifySmsResponse_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/VerifySmsResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifySmsRequest": {
				"properties": {
					"phoneNumber": {
						"type": "string",
						"description": "인증번호를 받은 휴대폰 번호",
						"example": "01012345678",
						"pattern": "^01[016789]-?\\d{3,4}-?\\d{4}$",
						"minLength": 10,
						"maxLength": 13
					},
					"code": {
						"type": "string",
						"description": "인증 코드 6자리 숫자",
						"example": "123456",
						"pattern": "^[0-9]{6}$",
						"minLength": 6,
						"maxLength": 6
					}
				},
				"required": [
					"phoneNumber",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"KakaoSignupResponse": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"signup"
						],
						"nullable": false
					},
					"user": {
						"properties": {
							"role": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"kakaoId": {
								"type": "string"
							}
						},
						"required": [
							"role",
							"email",
							"kakaoId"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"type": "string",
				"enum": [
					"user",
					"reformer"
				]
			},
			"AuthStatus": {
				"type": "string",
				"enum": [
					"PENDING",
					"APPROVED",
					"REJECTED"
				]
			},
			"KakaoLoginResponse": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"login"
						],
						"nullable": false
					},
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					},
					"user": {
						"properties": {
							"auth_status": {
								"$ref": "#/components/schemas/AuthStatus"
							},
							"role": {
								"$ref": "#/components/schemas/Role"
							},
							"email": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"role",
							"email",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"accessToken",
					"refreshToken",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"KakaoAuthResponse": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/KakaoSignupResponse"
					},
					{
						"$ref": "#/components/schemas/KakaoLoginResponse"
					}
				]
			},
			"TsoaResponse_KakaoAuthResponse_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/KakaoAuthResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LogoutResponse": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"statusCode",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TsoaResponse_LogoutResponse_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/LogoutResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignupResponse": {
				"properties": {
					"user": {
						"properties": {
							"role": {
								"$ref": "#/components/schemas/Role"
							},
							"nickname": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"role",
							"nickname",
							"email",
							"id"
						],
						"type": "object"
					},
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TsoaResponse_UserSignupResponse_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/UserSignupResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegistrationType": {
				"type": "string",
				"enum": [
					"LOCAL",
					"KAKAO"
				]
			},
			"UserSignupRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"registration_type": {
						"$ref": "#/components/schemas/RegistrationType"
					},
					"oauthId": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"over14YearsOld": {
						"type": "boolean"
					},
					"termsOfService": {
						"type": "boolean"
					},
					"privacyPolicy": {
						"type": "boolean"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"name",
					"email",
					"nickname",
					"phoneNumber",
					"registration_type",
					"over14YearsOld",
					"termsOfService",
					"privacyPolicy",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReformerSignupResponse": {
				"properties": {
					"user": {
						"properties": {
							"auth_status": {
								"$ref": "#/components/schemas/AuthStatus"
							},
							"role": {
								"$ref": "#/components/schemas/Role"
							},
							"nickname": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"auth_status",
							"role",
							"nickname",
							"email",
							"id"
						],
						"type": "object"
					},
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TsoaResponse_ReformerSignupResponse_": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/ReformerSignupResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "myform_reform",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/project": {
			"get": {
				"operationId": "SetTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TsoaResponse_string_"
								}
							}
						}
					}
				},
				"tags": [
					"project node CRUD"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/chat/rooms": {
			"post": {
				"operationId": "CreateChatRoom",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TsoaResponse_SimplePostResponseDTO_"
								}
							}
						}
					}
				},
				"summary": "채팅방 생성",
				"tags": [
					"채팅 기능"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"request": {
										"$ref": "#/components/schemas/CreateChatRoomDTO"
									}
								},
								"required": [
									"request"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/chat/rooms/list/{myId}": {
			"get": {
				"operationId": "GetChatRooms",
				"responses": {
					"200": {
						"description": "채팅방 목록 배열",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TsoaResponse_ChatRoomListDTO_"
								}
							}
						}
					}
				},
				"summary": "채팅방 목록 조회",
				"tags": [
					"채팅 기능"
				],
				"security": [],
				"parameters": [
					{
						"description": "채팅방 목록을 조회할 사용자 고유 아이디(테스트용)",
						"in": "path",
						"name": "myId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "채팅방 목록 필터 타입\n- 없음 : 전체 조회\n- INQUIRY: 문의 채팅방\n- ORDER: 주문제작 채팅방\n- UNREAD: 안 읽은 메시지가 있는 채팅방",
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ChatRoomFilter"
						}
					},
					{
						"description": "커서 기반 페이지네이션을 위한 커서 값, 마지막으로 조회된 채팅방의 ID",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "사용자 유형(테스트용)",
						"in": "query",
						"name": "myType",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"owner",
								"requester"
							]
						}
					}
				]
			}
		},
		"/auth/sms/send": {
			"post": {
				"operationId": "SendSms",
				"responses": {
					"200": {
						"description": "SMS 전송 완료",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TsoaResponse_SendSmsResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"success": {
												"statusCode": 200,
												"message": "SMS 전송이 완료되었습니다."
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "전화번호 형식 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "인증 시도 횟수 초과",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "입력한 휴대폰 번호로 인증 코드를 발송합니다.",
				"summary": "입력한 휴대폰 번호로 인증 코드를 발송합니다.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendSmsRequest"
							}
						}
					}
				}
			}
		},
		"/auth/sms/verify": {
			"post": {
				"operationId": "VerifySms",
				"responses": {
					"200": {
						"description": "인증 코드 검증 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TsoaResponse_VerifySmsResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"success": {
												"statusCode": 200,
												"message": "인증이 성공적으로 완료되었습니다."
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "인증 코드 불일치 및 형식 오류, 만료 또는 부재",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "인증 시도 횟수 초과",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "입력한 휴대폰 번호와 인증 코드를 검증합니다.",
				"summary": "인증 코드를 검증합니다.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifySmsRequest"
							}
						}
					}
				}
			}
		},
		"/auth/kakao": {
			"get": {
				"operationId": "LoginWithKakao",
				"responses": {
					"302": {
						"description": "카카오 로그인 페이지로 리다이렉트",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"summary": "카카오 로그인 시작 (로그인 페이지로 리다이렉트)",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"description": "로그인 모드 (user: 일반, reformer: 리폼러)",
						"in": "query",
						"name": "mode",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"user",
								"reformer"
							]
						}
					}
				]
			}
		},
		"/auth/kakao/callback": {
			"get": {
				"operationId": "KakaoCallback",
				"responses": {
					"200": {
						"description": "카카오 로그인 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/KakaoAuthResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"success": {
												"status": "login",
												"accessToken": "accessToken",
												"refreshToken": "refreshToken",
												"user": {
													"id": "userId",
													"email": "userEmail",
													"role": "reformer",
													"auth_status": "PENDING"
												}
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "입력한 mode의 값이 유효하지 않습닏..",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "카카오 인증에 성공했으나 유저 정보를 가져오지 못했습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"summary": "카카오 로그인 콜백 (카카오 인증 후 회원 정보 조회)",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "로그아웃 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TsoaResponse_LogoutResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"success": {
												"statusCode": 200,
												"message": "로그아웃이 성공적으로 완료되었습니다."
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "로그인 정보를 찾을 수 없습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/signup/user": {
			"post": {
				"operationId": "SignupUser",
				"responses": {
					"201": {
						"description": "일반 회원가입 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TsoaResponse_UserSignupResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"success": {
												"user": {
													"id": "userId",
													"email": "userEmail",
													"nickname": "userNickname",
													"role": "user"
												},
												"accessToken": "accessToken",
												"refreshToken": "refreshToken"
											}
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignupRequest"
							}
						}
					}
				}
			}
		},
		"/auth/signup/reformer": {
			"post": {
				"operationId": "SignupReformer",
				"responses": {
					"201": {
						"description": "리폼러 회원가입 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TsoaResponse_ReformerSignupResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"success": {
												"user": {
													"id": "userId",
													"email": "userEmail",
													"nickname": "userNickname",
													"role": "reformer",
													"auth_status": "PENDING"
												},
												"accessToken": "accessToken",
												"refreshToken": "refreshToken"
											}
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"data": {
										"type": "string"
									},
									"portfolios": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"data",
									"portfolios"
								]
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}